Create render folder
